+++ server/standalone/migration/000010_add_timestamp_helper.down.sql
@@ -1,0 +1,1 @@
+DROP FUNCTION IF EXISTS to_rfc_3339(timestamptz);